Fix input systems configuration
No model change needed with this logic, only viewmodel changes. In the
viewmodel, hasCustomInputSystemsOverride becomes an InputSystemSettings
type so that it has a set of role and group booleans instead of being
a single boolean, and it is used to keep track of whether the user has
ever changed any of the checkboxes. If so, an override is automatically
assumed and set as such in the config. But there's also a Reset button
which lets the user go back to a "No override" status.


Former-commit-id: 78eb880ca0edf6bc2bb0497996dc56696bfd79f7