Fix activity log not showing in Safari (#706)
* Fix activity log not showing in Safari

Safari parses timestamps strictly according to RFC 3339, which
requires a colon in the timezone ("+07:00", not "+0700"). The
default ISO8601 format codes in PHP use "+0700" format, so we
need to change how we encode dates in JSON so that Safari (and
other strictly-compliant browsers, like Safari Mobile) accept
the dates the server returns.

* Adjust unit tests for new expected format

The expected value of JSON-encoded dates should now include a colon
in the time zone format, so the two tests that expected the old,
colonless format need to be adjusted to expect what the code now does.