Fixes DateTime inequality due to PHP bug
In PHP 7.0, the following line was not running correctly in MongoDecoder.php:
$model->setTimestamp($newDateTime->getTimestamp());
The model's timestamp was not being overwritten and its previous value persisted.
Unforunately this is what the test was expecting and thus it passed.
This bug was fixed as of PHP 7.3 but the way we are converting dates prevents preservation of milliseconds.
Thus in order to pass the test, we must truncate milliseconds.

remove comment


Former-commit-id: 3deeccd3c568bc579bfa1ee49176f527299470a4