Rendered entries now show correct part of speech
Previously the optionLists data wasn't making it all the way to the
dc-rendered component, so the component couldn't look up proper
abbreviations for parts of speech and had to resort to using the key
instead. This was mostly invisible for projects that had an English
language user interface, but for non-English interfaces the English
part-of-speech abbreviations were being wrongly shown. Fixed.

This should fix https://jira.sil.org/browse/LF-319