Force angular-* modules to match angular version
This should solve the "lowercase is not a function" error that was
sometimes showing up. The cause is that between Angular 1.6 and 1.7,
they removed the long-deprecated angular.lowercase function, which
some code (including angular-sanitize) had been depending on. This
means that the 1.6 to 1.7 upgrade is a breaking change, which means
that it's not safe to use a package version like "^1.6.10" for the
angular-sanitize module, because that will pull in 1.7.x (since npm
thinks that angular is following SemVer). The angular-* modules need
to be the same version as angular to be safe, so we use "1.6.10"
instead of "^1.6.10" in the angular-sanitize version in package.json.