Fix sort and filter handling of input systems and fields
Previously settings for which users should be able to see fields and
input systems were not respected by the filter and sort. This is because
a client-side modification to the config has to happen before it
accurately indicates what the user can see. This commit makes the sort
and filter logic use that modified config.