Fix blank comments panel when coming from entry list
If the user navigated back to the entry list while having the comments
panel (or the recent-activity panel) open, coming back from the entry
list to an entry would leave a blank space where the comments panel is
supposed to be, and it would be impossible to show the comments panel by
clicking on the Comments button.

Fixes LF-262.