Fix bug where entryListModifiers.sortBy is null
It is often happening that there's an invalid sortBy parameter in the
URL, and our code currently assumes that the URL parameter will always
be found in the sort options; when that wasn't happening, the
findSelectedFilter function was returning null or undefined. Now we
check the result first, and only assign it to entryListModifiers.sortBy
if it is valid.

This passes E2E tests once PR #768 has been merged.