Fix E2E tests of projects page
This is not a good long-term fix since it exercises an app (the user
management app) which we no longer expose to the user in our UI and so
we'll probably remove it at some point. But in the short term, it gets
the E2E tests passing again. In the long term, we should change the
addUserToProject implementation to use the user's email address and
enter it in the new "Share With Others" dialog.