WIP fixing LF-372: parse inline LIFT ranges right
LiftImport needs to be able to cope with a <range> element that doesn't
have an href attribute, but defines the range inline. Our current code
had an assumption that there will always be an href attribute, but the
LIFT spec (as defined by the RelaxNG schema) lists href as optional. The
code already knows how to parse <range> elements defined inline, we just
need to make sure we don't assume the presence of an optional attribute.