Restart remote PHP by creating trigger file
Instead of doing "ssh sudo service php7.3-fpm restart", we now create a
pair of systemd unit files that watch for a specific filename, and
restart the php7.3-fpm service when that filename appears. Then the Gulp
task just has to create that file via rsync, and no sudo permissions are
needed on the server.