Fix blank project page for Tech Support role
Many projects don't have anything configured for the Tech Support role,
since it was newly added, so those projects end up throwing an exception
in the filter() call in removeDisabledConfigFields. And since the filter
didn't return true, all fields end up being stripped out. The Tech
Support role is supposed to have the same rights as a project manager,
so if it's looked for and not found, we give manager rights.