Make user list robust against missing projects
When the site admin tries to list users, if there is even one user who
was in a deleted project but somehow that project was not removed from
their membership list, the user list fails to load. This change makes it
just ignore that missing project when loading the user list.