Add ability to call Language Depot API
Currently calls LD API via localhost:4200, which is where the LD API's
dev server is configured to run the service. Eventually we'll change the
URL called to be admin.languagedepot.org and create a Webpack config in
Language Forge to redirect that URL to the dev server; that way the code
checked into Git can have the correct URL in it.