Better handle error results from LD API calls
By default, GuzzleHttp throws PHP exceptions when it gets a result in
the 400 or 500 range. We don't want that, because when it throws an
exception it truncates the message that the API returned, and we want
the entire message. So we tell GuzzleHttp not to throw exceptions, and
we throw them ourselves (which JsonRpcServer catches and returns as an
"ok=false" result). That way .NET stacktraces are preserved.