Ensure projects and users grid render separately
The ui-grid library users the grid element's ID to keep track of it; if
two grids on the same page aren't given separate IDs, then one will end
up blank. Adding IDs fixes it so that now, whichever one of the LD
Projects and LD Users tab you click on second will render properly.
(Before, the first tab you clicked on would render but the second one
would be blank because ui-grid didn't realize it was a different grid).