Now user projects view shows LD projects
Improvements needed: Need to discuss what to call these in this view.
Right now we call them "Language Depot" projects, but is that going to
be confusing to people? The distinction between a dictionary project and
a Language Depot project? Should this be integrated into this view, or
separated out to a different view?