Input system tag, not name, in comment.regarding
We've had this bug for several years, so a migration script may be
required. But some parts of the code assume that a comment's "regarding"
object will have inputSystem be the *tag* ("en"), and yet it's being set
to the *name* ("English"). We never use it to display the name to the
user (that's what inputSystemAbbreviation is for), so this changes it to
actually be the tag that the rest of the code is expecting.