Fix various issues with ansible install script (#813)
* add missing base php apt package

* ensure system is using PHP 7.3

* explicitly set PHP version in all PHP package names

* address review comments

- remove redundant header
- use {php_version} in update-alternatives command

* add php7.3-xml dependency

The php-pear package requires php7.4 but unfortunately the pear command cannot run in the 7.3 environment unless php7.3-xml is explicitly installed.  Added this package to fix the issue with pear not running properly under v7.3

* remove "refreshDeps" from ansible, replace with "gulp dev-build"

the refreshDeps.sh shell script is overkill in this scenario as it runs a fresh 'npm install' and 'composer' install before doing a dev-build.  When run as an ansible script, these tasks are run a few steps before and don't need to be run again.

* update npm to latest version

The ansible script is failing on npm install inside of ansible, but works on the command line.  Updating npm to the latest version fixes this issue.