wip node container (not working yet)
Thought I could get a node container working.  This is a start but so far the volume mapping is overwriting the node_modules created in the image, so I'm not sure what I'm going to do about that at the moment.  I thought .dockerignore would help me here but it's either not working or it's not configured correctly