php-test Dockerfile: inherit from php-api
I used the image: property in the docker-compose file to name the image "php-api:latest" and then reference that in the FROM statement of the php-test Dockerfile

Unfortunately, there are loads of PHP tests failing now because of the
changes I made to the website instances.  I'll address those failures in
a future commit.