"languageforge.localhost" -> "localhost" find/replace
A find/replace that changes the domain to "localhost" for all relevant source code files.  I didn't bother changing some files that are no longer relevant or I wanted to preserve for historical purposes e.g. ansible files.

Not all of the files changed are actually relevant anymore.  There are a number of protractor config files, and we probably only need one at this point.