e2e tests running in docker
Refactor e2e tests and create specific e2e containers to run in docker environment.

This creates 3 new containers:
e2e-api
e2e-ui
selenium

The new api and ui containers contain test-specific files and setup that are not appropriate for the existing api and ui images.  The selenium container runs the tests in a chrome browser.