Move to server-side JWT for LD API access
This allows us to move the LD API security into the Language Depot API
server where it belongs. The user logged in to Language Forge will be
looked up by email address in the Language Depot server; as long as the
email addresses match (and are unique), a JWT will be created that will
tell the Language Depot API server who's trying to access the method.