Fix reset password part of E2E traversal test (#837)
* Fix reset password part of E2E traversal test

The failure on the reset password part of the E2E traversal test is
because the traversal test is designed to be run first, before anything
else, but the test locator logic is running tests in subdirectories
before it runs tests in the parent directory, so that the
test/app/bellows/bellows-traversal.e2e-spec.ts test is running *after*
the test/app/bellows/reset-forgotten-password.e2e-spec.ts test. The
bellows-traversal test needs to run first. Moving the
reset-forgotten-password test into the parent directory fixes this issue.