Fix extra params in e2e-ui/run.sh (#843)
* Fix extra params in e2e-ui/run.sh

The run.sh file in docker/e2e-ui was originally designed to pass in the
name of the site (sf for Scripture Forge, lf for Language Forge, etc) as
its first parameter, so it was designed to swallow its first parameter
and pass any later parameters on to the Gulp task, for things like
--specs=change-password. Now that we're no longer calling it with the
site name, we no longer need to strip the first parameter.

Also added quotation marks around the `$@` usage so that spaces won't be
dropped: see https://www.grymoire.com/Unix/Sh.html#uh-42 for why.