Don't let Docker cache too aggressively
Docker's cache system assumes that RUN commands are deterministic: with
the same input they will produce the same output. Therefore, if the
previous line was taken from Docker's cache, the "RUN git clone" line
will also have its output cached and we won't get any new Git commits.
The Docker ADD command knows about remote URLs and HTTP caching, so if
we use ADD to download a GitHub tarball we won't have any issues: it
will check the relevant HTTP headers like Etag to verify whether it
already knows the contents of the URL, and if the contents haven't
changed it will skip the download (and the rest of the build will be
from Docker cache as well).