Fix project config E2E tests (#855)
* Fix project config E2E tests

The "Select All" tests were failing because the wrong Angular directive
was being used for the "Select All" checkboxes in the field config. The
ng-click directive fires before ng-model, so at the time the ng-click
directive was firing on a "Select All" checkbox, the model was holding
the *old* value for the checkbox, not the *new* value. Result: when you
turned a "Select All" box *on*, it would turn all the checkboxes in its
row and column *off*.

Replacing ng-click with ng-change on all the Select All checkboxes
solved the issue. Since ng-change runs *after* ng-model, by the time the
handler runs and updates the state of the checkboxes in the row and
column, the model contains the *new* value and the "Select All" feature
works like you'd expect it to work.