fix failing LFMerge PHP tests
Interestingly, the PHP tests were failing because the tests assumed that nothing was running under PID 1, probably a fair assumption in the former environment.  In Docker, the CMD is always run as PID 1, so it is always present.  Changing the PID to 2 (where nothing is running) makes the two tests pass.  These tests are pretty low value and don't actually test the presence of LFMerge.