remove apache systemd service file
@rmunn and I had a discussion about how Docker doesn't run systemd and therefore this file addition is actually useless/harmless.

This file was originally included by myself and @longrunningprocess in order to address the PrivateTmp=false requirement that is stated in our Ansible documentation.  Upon closer inspection, PrivateTmp is a construct of systemd itself (tmp folder isolation per service) and doesn't exist when running apache-foreground inside Docker.  Fortunately, we don't want a PrivateTmp isolation (we have a documented bug in production involving PrivateTmp isolation) and so going with Docker this issue no longer exists.

Solution: delete the file and Dockerfile reference.