Feature/LFMerge in api container (#865)
* move lfmerge install into api container

I did some tests and amazingly, the lfmerge deb package installs in debian buster with no modifications.  This package was designed to be installed in ubuntu bionic.

* setup LFMerge permissions in api

* enable S/R E2E tests

Enable the tests that Robin had commented out earlier

* add back two LFMerge PHP unit tests

These tests were previously deleted, but now with LFMerge installed we can safely add them back.

* fix failing LFMerge PHP tests

Interestingly, the PHP tests were failing because the tests assumed that nothing was running under PID 1, probably a fair assumption in the former environment.  In Docker, the CMD is always run as PID 1, so it is always present.  Changing the PID to 2 (where nothing is running) makes the two tests pass.  These tests are pretty low value and don't actually test the presence of LFMerge.

* address review comments

This removes a copy/paste oversight.  No actual reason to remove/purge 'curl'

* remove apache systemd service file

@rmunn and I had a discussion about how Docker doesn't run systemd and therefore this file addition is actually useless/harmless.

This file was originally included by myself and @longrunningprocess in order to address the PrivateTmp=false requirement that is stated in our Ansible documentation.  Upon closer inspection, PrivateTmp is a construct of systemd itself (tmp folder isolation per service) and doesn't exist when running apache-foreground inside Docker.  Fortunately, we don't want a PrivateTmp isolation (we have a documented bug in production involving PrivateTmp isolation) and so going with Docker this issue no longer exists.

Solution: delete the file and Dockerfile reference.