Remove Scripture Forge code and references (#873)
* remove scripture forge src files

* remove scripture forge from WebsiteInstances.json

* remove references to scripture forge in README

* remove scripture forge references in various src files

These edits are untested and very likely to not get correct on the first round.

* remove retired files and semdomtrans files

* remove browserStack conf files

* teardownTestEnvironment is not necessary in Docker

* cleanup random files

* remove cross-site login from AuthUserProvider.php

* slash and burn in README.md

* delete refreshDeps.sh

* remove unused references

* remove SF API methods from sf.php

* remove references

* remove special cases and references to sfchecks

* updated recommended VSCode extension to ESLint

TSLint extension has been depricated in favor of ESLint

* fix up PHP tests (continue removing sfchecks / semdomtrans / scriptureforge)

all PHP tests are now passing after removal of sfchecks / scriptureforge specific components

* php intellisense vscode plugin tweak

The php intellisense vscode plugin recommends turning off basic PHP suggestions so that they don't interfere with the plugin's suggestions.

* volume map src files in test-api service

This is helpful when debugging and fixing up PHP tests

* delete src/vendor folder

I'm not sure exactly how this got checked in, but I think it needs to be removed in order to build

* fix E2E tests

remove references to linkService.text and linkService.question which no longer exist.

* fix src volume map in test-api to get PHP tests running

I naively added a volume map for the src folder forgetting that I cannot do that.  At least because of conflicts with the vendor folder.  This change should provide most of what I need and also get the PHP tests running