Add an hgweb service for viewing LD projects
This will eventually be part of the Language Depot deployment, not
Language Forge. But until the Language Depot containerization is
working, having this container as part of the Language Forge set allows
development of the Language Forge UI for viewing repositories on
Language Depot. The ld-hgweb container can be used as a sample of what
repositories on Language Depot will look like when viewed through the
Language Forge interface.