add test-specific `docker-compose build`
This was undone in previous commits as a simplification step, however that simplification relied on the `docker-compose run` command building the necessary image.  The problem with that is that it does not "rebuild if necessary" when a change has been introduced.  This causes the run command to use older image versions if they exist.

This change ensures that the test-specific images get built every time.  Without this commit the build server fails due to using a previous image that does not have this branch's changes.