get XDebug 3 working in Docker
These changes enable XDebug PHP debugging using VSCode + PHP Debug extension.

I used a number of instructions/tidbits from the internet to figure out how to get this working.  Most of the documentation on the internet refers to XDebug 2 configurations, however we are running XDebug v3 which has different configuration options.

In addition, for this to work, the XDebug programming running inside Docker needs to connect to VSCode on the host on port 9003.  Normally services running inside Docker don't need the ability to connect to services running on the host however XDebug is a different kind of animal and needs this ability.

I added the `extra_hosts` in docker-compose to make this possible.  This requires Docker v20.10 or later to use the special `host.docker.internal`

Some additional notes on debugging PHP using VSCode and XDebug:

- Because PHP composer dependencies only exist inside the docker container and not in your source tree, XDebug cannot step into vendor/composer dependencies (unless you decide to download dependencies on your host).  If you get a VSCode error stating that it cannot find a file, check to make sure that it is not in the vendor folder.

VSCode allows you to break on 'everything'.  I suggest turning this off so that the debugger doesn't break inside composer dependencies by default.