Make Website::baseUrl() use WEBSITE env var
This ends up being the smallest possible change from master that will
make OAuth and Communicate.php methods work. At some point, probably
when we remove the Website class, we'll also need to change the places
that call `$website->baseUrl()`, but I've chosen not to do that in this
commit so that it will interfere as little as possible with the
chore/removeWebsiteClass branch.