Set WEBSITE env var during TeamCity unit tests
The WEBSITE environment variable is now required to run PHP unit tests
successfully, and we can't just set it in TeamCity build properties
because variables from the environment aren't automatically passed
through by `docker-compose run`. So when running the unit tests on
TeamCity, we need to set the required env var this way.