add shell scripts to document kubectl SSH tunnel
These shell scripts are intended to be documentation for how I setup a SSH tunnel to both the APP and Mongo containers for the purposes of using rsync to sync files between US containers (as opposed to bouncing files off Thailand between containers).

The process was educational and fun to see working.

These can probably be discarded after we complete our migration tasks and have a new production environment.