Better handling of field input systems overrides
In project config, field input systems are supposed to override user
input systems, not the other way around. Previously the logic was
incorrectly having user input systems override field input systems.

This fix changes the logic for input systems so that the list of input
systems for the field and the list of input systems for the user are
ANDed together to get the override list. I.e., if the user has input
systems A, B and D selected but C turned off, and the field is supposed
to show input systems B, C and D, then the user will see just input
systems B and D for that field.

We also get rid of the Reset buttons on the input systems settings. It
was insufficiently obvious to users what they did, and with the new
logic for calculating field overrides they're not needed any longer.