E2E regression tests to catch this in the future
The E2E tests turned up one feature of the config view where the code
was working as intended, but was causing the E2E tests to fail because
the intent of the code was not something I expected. Specifically, when
the input system config is populated, we called the isInputSystemUsed
function and if it returned false, we unchecked that input system in the
config. The intent was for projects that had extraneous input systems in
their config (left over from a bug in early Language Forge where the
Thai input system would be added to projects no matter what language
they were intended for), those input systems would be automatically
ignored. But in my E2E tests, I added the Thai input system, used only
by the Word (lexeme) field. Then when I unchecked the Thai input system
for the lexeme field, the code detected that it was no longer being used
and automatically unchecked it from the input systems config for all
four roles.

Now that the feature is in place where input systems are ANDed together
between role config and field config, that feature (of unchecking unused
input systems) is no longer desired, so I'm removing it with this
commit. This makes the user experience easier to understand, *and* makes
the E2E tests pass.