Better handling of field input systems overrides (#997)
* Better handling of field input systems overrides

In project config, field input systems are supposed to override user
input systems, not the other way around. Previously the logic was
incorrectly having user input systems override field input systems.

This fix changes the logic for input systems so that the list of input
systems for the field and the list of input systems for the user are
ANDed together to get the override list. I.e., if the user has input
systems A, B and D selected but C turned off, and the field is supposed
to show input systems B, C and D, then the user will see just input
systems B and D for that field.

We also get rid of the Reset buttons on the input systems settings. It
was insufficiently obvious to users what they did, and with the new
logic for calculating field overrides they're not needed any longer.

* Fix E2E test

One E2E test is affected by this change: it used to use the Reset
buttons to reset state changed by previous tests, but now we need to
reset that state in a more explicit way to make the test actually test
what it's supposed to be testing.

* E2E regression tests to catch this in the future

The E2E tests turned up one feature of the config view where the code
was working as intended, but was causing the E2E tests to fail because
the intent of the code was not something I expected. Specifically, when
the input system config is populated, we called the isInputSystemUsed
function and if it returned false, we unchecked that input system in the
config. The intent was for projects that had extraneous input systems in
their config, those input systems would be automatically ignored. But in
my E2E tests, I added the Thai input system, used only by the lexeme
field. Then when I unchecked the Thai input system for the lexeme field,
the code detected that it was no longer being used and automatically
unchecked it from the input systems config for all four roles.

Now that the feature is in place where input systems are ANDed together
between role config and field config, that feature (of unchecking unused
input systems) is no longer desired, so I'm removing it with this
commit. This makes the user experience easier to understand, *and* makes
the E2E tests pass.