Better setup for E2E test container (#999)
This ensures that the E2E test container will be ready for debugging as
soon as it starts up. (Previously, VS Code could attach to the container
before it was actually ready, and it wasn't always clear why you had to
wait a minute before pressing the "Launch Debug Session" button).