Fix failing E2E tests
This ensures that tests which modify state just for their own sake
(i.e., no other tests are depending on the modified state later) will
restore the state to how it was after they run.

With this change, we can now run the config-fields.e2e-spec tests
multiple times in a row, passing each time.