Optimize required fields list
The required fields list is not something that will need to be
constructed dynamically. It may change from time to time as we develop
the site, but it's not something that needs to be calculated every time
the config page is rendered, let alone something that needs to be
calculated every time the site decides how to render a checkbox. A
static array will ensure that the array is only created once, rather
than creating and throwing an array once per checkbox rendering
decision. If we need to implement logic in the required fields list in
the future, we can turn it back into a function at that time, and take
the corresponding efficiency hit when we have to.