Update logic to include roles, and simplify
Since we're disabling the checkboxes for required fields in both the
users and roles columns, we can simplify the logic by not having to deal
with two different rules for roles and users. This also fixes a corner
case where it was possible to turn off the lexeme field checkbox in a
user column by first turning off one of the roles, then using the Select
All checkbox for the user column. Now the Select All checkboxes for
columns properly ignore required fields and force them to be on.

This causes four E2E tests to fail, because they were expecting the
Select All columns for roles to be able to turn off the lexeme field.
Next commit will fix the four failing tests.