Update E2E test logic so tests pass now
Changed the logic for isAllCheckboxes to ignore any checkboxes that are
disabled. So the function actually becomes isAllEnabledCheckboxes,
though I chose not to rename it because that's not a great name and I
haven't yet come up with an actually good one. We might rename that
function in the future. But in the meantime, by skipping any disabled
checkboxes when verifying the effects of a Select All, this ensures that
the E2E tests have the desired logic. And now they all pass.