prevent user from unchecking "word" in project dictionary configuration #938 (#944)
* prevent user from unchecking "word" in dictionary configuration #938

* Update src/angular-app/languageforge/lexicon/settings/configuration/configuration-fields.component.ts

Co-authored-by: Christopher Hirt <chris@hirtfamily.net>

* Update src/angular-app/languageforge/lexicon/settings/configuration/field-unified-view.model.ts

Co-authored-by: Christopher Hirt <chris@hirtfamily.net>

* Update src/angular-app/languageforge/lexicon/settings/configuration/field-unified-view.model.ts

Co-authored-by: Christopher Hirt <chris@hirtfamily.net>

* Update src/angular-app/languageforge/lexicon/settings/configuration/field-unified-view.model.ts

Co-authored-by: Christopher Hirt <chris@hirtfamily.net>

* Update src/angular-app/languageforge/lexicon/settings/configuration/field-unified-view.model.ts

Co-authored-by: Christopher Hirt <chris@hirtfamily.net>

* Update src/angular-app/languageforge/lexicon/settings/configuration/field-unified-view.model.ts

Co-authored-by: Christopher Hirt <chris@hirtfamily.net>

* Update src/angular-app/languageforge/lexicon/settings/configuration/field-unified-view.model.ts

Co-authored-by: Christopher Hirt <chris@hirtfamily.net>

* Refactor requiredConfigurationFields to ensureRequiredConfigurationFields

* Added fieldName to BaseSettings as it is needed by both FieldSettings and InputSettings classes.
Commented e2e tests for selectAll per stand up discussion.

* Changes e2e tests to expect all configuration checkbox to be checked.

* e2e still failing on "select all" changing false to true

* e2e still failing on "select all" changing false to true

* e2e still failing on "select all" changing false to true

* Uncomment failing E2E tests

* Fix failing E2E tests

This ensures that tests which modify state just for their own sake
(i.e., no other tests are depending on the modified state later) will
restore the state to how it was after they run.

With this change, we can now run the config-fields.e2e-spec tests
multiple times in a row, passing each time.

* Optimize required fields list

The required fields list is not something that will need to be
constructed dynamically. It may change from time to time as we develop
the site, but it's not something that needs to be calculated every time
the config page is rendered, let alone something that needs to be
calculated every time the site decides how to render a checkbox. A
static array will ensure that the array is only created once, rather
than creating and throwing an array once per checkbox rendering
decision. If we need to implement logic in the required fields list in
the future, we can turn it back into a function at that time, and take
the corresponding efficiency hit when we have to.

* Update logic to include roles, and simplify

Since we're disabling the checkboxes for required fields in both the
users and roles columns, we can simplify the logic by not having to deal
with two different rules for roles and users. This also fixes a corner
case where it was possible to turn off the lexeme field checkbox in a
user column by first turning off one of the roles, then using the Select
All checkbox for the user column. Now the Select All checkboxes for
columns properly ignore required fields and force them to be on.

This causes four E2E tests to fail, because they were expecting the
Select All columns for roles to be able to turn off the lexeme field.
Next commit will fix the four failing tests.

* Update E2E test logic so tests pass now

Changed the logic for isAllCheckboxes to ignore any checkboxes that are
disabled. So the function actually becomes isAllEnabledCheckboxes,
though I chose not to rename it because that's not a great name and I
haven't yet come up with an actually good one. We might rename that
function in the future. But in the meantime, by skipping any disabled
checkboxes when verifying the effects of a Select All, this ensures that
the E2E tests have the desired logic. And now they all pass.

Co-authored-by: Christopher Hirt <chris@hirtfamily.net>
Co-authored-by: Robin Munn <rmunn@pobox.com>