Run lfmerge queue manager on container startup
This will allow us to fix any send/receive sessions that were partway
through being run when the previous container was killed (e.g. if the
pod was moved to a different Kubernetes node during a S/R process). We
use `which lfmergeqm && ...` to ensure that this doesn't cause the
container to fail if the entrypoint script ends up run on a container
without lfmerge (which might happen once we separate lfmerge into its
own container in the future).