Remove a lot of now-unused code
Now that we don't show the Language Depot projects in the LF projects
list view, a lot of the ldproject code is unused. The only code that's
actually used is the ldproject-members component, which we use in the
siteadmin view. In theory I could move that back into site admin, but
I prefer to leave it in the ldproject app directory so that if we decide
to bring back the ldproject app in the future (with a redesign) then it
will involve fewer code changes.