Run lfmergeqm every 24 hours (#1034)
This will allow lfmergeqm's Janitor class to find and clean up any
failed syncs that aren't in the queue (state is SYNCING but there's no
active lfmerge process handling that project). The main reason this can
happen is if the app container was restarted in the middle of a
Send/Receive, killing a previous lfmerge process.