Merge app-base into app Dockerfile; establish BUILD_VERSION in GA (#1030)
The app-base Dockerfile was created for developer convenience to
simplify and speed up building of the docker image during development of
the Dockerfile.

In preparation for a production image, we can merge these two back
together and provide a means for switching between production and
development images via a build argument "environment".

This gets us started with a build version stamp generated by GA and
inserted into the built image. The build number is 3 places, e.g. 9.9.9.
The first two are taken from the version_number.txt file at the repo
root, and the third will come from the run number of the GitHub Action
that pushes the release.

Commits included in this PR:

* merge app-base into app Dockerfile
* establish BUILD_VERSION variable in GA
* Update version.php with BUILD_VERSION env var
* Move lfmergeqm-background script file
* address review comments

Co-authored-by: Robin Munn <rmunn@pobox.com>