Ensure E2E test consistency
By restarting the app-for-e2e container each time we run `make
e2e-tests`, we can ensure that the E2E tests run consistently even if a
previous run made a state change in the database and didn't clean it up
(e.g., if a test failed halfway through and didn't reach the code
intended to reset the state change). The app-for-e2e container runs the
setupTestEnvironment.php script on startup, which initializes the
database with the E2E test data. So simply restarting it works in nearly
every case, except after a `make clean` when the container doesn't yet
exist. In that case, `docker-compose up -d app-for-e2e` starts it up.