Fix `make dev` (#1050)
* Fix `make dev`

The ui-builder changes from #1038 accidentally got reverted in #1040,
probably due to a rebase onto a master branch that hadn't yet had the
merge of #1038 applied yet. This should fix the issue.

* Build ui-builder during E2E tests

This isn't actually needed long term, and at some point we'll adjust the
E2E test Dockerfile to pull from something else. But for now it's needed.