Fix mistake in entrypoint.sh script (#1059)
* Fix mistake in entrypoint.sh script

It's not `if [ which some-cmd ]`, it's `if which some-cmd`. It can also
be a good idea to redirect the output of `which` to `/dev/null` so that
you don't end up printing the full path of the command in your server
logs (which is almost never necessary).

* Fix "activation of module imklog failed" errors

Inside Docker containers, rsyslog isn't allowed to run the imklog
module, which reads kernel messages. Since kernel messages aren't
needed (or wanted) in Docker container logs, just comment out the
line in /etc/rsyslog.conf that loads imklog.