Tech Support user can manage project members
The intent of the Tech Support user role is that it has the same rights
as a project manager, i.e. it can manage all aspects of the project. The
"Share With Others" component was not giving Tech Support users the same
rights as the manager: Tech Support was unable to list project members,
manage their roles, and invite new users. This fixes that omission.