add mongo dump and restore script via kubectl
This script is intended to be run from my (Chris's) machine, but can be adapted for other machines.  It documents as a script each step of dumping production mongo data and restoring it on the running db container.

Also modified the kubeSetup.sh to setup both app and db containers simultaneously