Send lfmerge logs to stdout in lf-app container
This will ensure that lfmerge logs get sent to stdout whether or not
https://github.com/sillsdev/LfMerge/pull/134 is merged. Currently
lfmerge runs in the lf-app container so its log messages end up mixed
together with the PHP logs in `kubectl logs` output, but it's pretty
easy to sort them out. If that mixing is not desired, then the
entrypoint.sh changes can be reverted and `kubectl exec` can be used to
run `tail -f /var/log/syslog` on the container instead.