Silence Dart Sass warnings with -q
Fixes #1152

Add `-q` flag to the `sass` command to suppress all warning messages.

The current deprecation warnings are related to our version of bootstrap, and we don't know how much work it would be to upgrade, so I decided the easiest thing to move forward is to silence those warnings.