Go directly to project on sign-in (#1140)
When users log in or sign up, if they have a most-recently-used project
or are a member of only one project, they are taken directly to that
project instead of seeing the "My Projects" list. However, the "My
Projects" breadcrumb and "List All Projects" menu entry still work and
take users to the projects list, so that anyone who's used to using
those links won't have to change habits.