establish appropriate capitalization hints for fields (#1156)
When mobile users were trying to add data in the app, the device keyboard, e.g., Android keyboard, would pop up and already be in a "title-case" mode where the first letter would be capitalized but this is a bit cumbersome for users because he/she would have to go back to the word and lowercase the first letter. So this request is made to help us give the user's browser some hints that the user's keyboard should not "autocapitalize" for given fields or perhaps the keyboard should "all caps" for other fields. It's worth noting that this is not something the application can necessarily force upon the user's keyboard but hopefully these "hints" will be respected in many, if not all, cases.