Adding or removing senses or examples is non-delta
Adding or removing a sense or example has to be a non-delta update so
that the existing code in LexSense, etc., can do what it needs to do to
fix up GUIDs. This is a requirement specific to our current PHP
implementation, but it will not be necessary once we switch to a
different backend.