fix CSS magic height numbers for list containers
The lexiconItemListContainer and compactEntryListContainer both contained non-optimal magic height numbers.

The new numbers provide a more consistent maximum height as well as no app surface scrollbar.

The only case where a scrollbar appears is when the search options dialog is exercised, and this is too complicated to fix without a more major CSS/HTML reshuffle.

Also remove line in FitTextDirective that fiddles with the compactListItemContainer height.  It should not.